+++ packages/docs/src/components/Content/ResponsiveExample.tsx
