--- packages/docs/src/components/ContentRenderer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ContentRenderer';
